requirements.txt
Flask==2.1.1
python-dotenv==0.20.0
pyodbc==4.0.32